+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Chain selection
+- Search bar to block leaves keys
 
 ## [1.1.3] - 2021-11-11